# apps/server/internal/docs/swagger.yaml
basePath: /
definitions:
...
  config.createConfigRequest:
    properties:
      configKey:
        type: string
      configName:
        type: string
      configType:
        type: string
      configValue:
        type: string
      remark:
        type: string
    required:
    - configKey
    - configName
    - configValue
    type: object
  config.updateConfigRequest:
    properties:
      configKey:
        type: string
      configName:
        type: string
      configType:
        type: string
      configValue:
        type: string
      remark:
        type: string
    type: object
...
  resp.Response:
    properties:
      code:
        type: integer
      data: {}
      error: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: Internal admin platform API documentation.
  title: Admin Service API
  version: 0.1.0
paths:
...
  /v1/monitor/cache:
    get:
      description: 查看 Redis 运行状态与内存信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 缓存概览
      tags:
      - Monitor/Cache
  /v1/monitor/cache/list:
    get:
      description: 按库、模式分页查询缓存键
      parameters:
      - description: 页码
        in: query
        name: pageNum
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: Key 模式
        in: query
        name: pattern
        type: string
      - description: Redis 库
        in: query
        name: db
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 缓存键列表
      tags:
      - Monitor/Cache
...
  /v1/system/configs:
    get:
      description: 按名称、Key、类型过滤配置
      parameters:
      - description: 参数名称
        in: query
        name: configName
        type: string
      - description: 参数Key
        in: query
        name: configKey
        type: string
      - description: 参数类型
        in: query
        name: configType
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 获取参数配置列表
      tags:
      - System/Config
    post:
      consumes:
      - application/json
      description: 创建一条系统配置
      parameters:
      - description: 配置参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/config.createConfigRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 新增参数配置
      tags:
      - System/Config
  /v1/system/configs/{id}:
    delete:
      description: 根据ID删除配置
      parameters:
      - description: 配置ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 删除参数配置
      tags:
      - System/Config
    get:
      description: 根据ID查询配置
      parameters:
      - description: 配置ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 获取参数配置详情
      tags:
      - System/Config
    put:
      consumes:
      - application/json
      description: 更新配置项
      parameters:
      - description: 配置ID
        in: path
        name: id
        required: true
        type: integer
      - description: 配置参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/config.updateConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 修改参数配置
      tags:
      - System/Config
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 输入 Bearer Token，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"